# apps/backend/public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Modfolio",
    "description": "Generate fast, beautiful and consistent embeddable cards and badges for Modrinth, CurseForge, Hangar and Spigot content.",
    "version": "unknown",
    "contact": {
      "name": "Creeperkatze",
      "url": "https://github.com/creeperkatze/modfolio"
    },
    "license": {
      "name": "AGPL-3.0",
      "url": "https://spdx.org/licenses/AGPL-3.0-only.html"
    }
  },
...
  "servers": [
    {
      "url": "https://modfolio.creeperkatze.de",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/modrinth/user/{username}": {
      "get": {
        "tags": ["Modrinth"],
        "summary": "Get Modrinth user card",
        "description": "Returns a rich SVG card showing user stats and top projects.",
        "parameters": [
          { "$ref": "#/components/parameters/Username" },
          { "$ref": "#/components/parameters/Color" },
          { "$ref": "#/components/parameters/BackgroundColor" },
          { "$ref": "#/components/parameters/ShowProjects" },
          { "$ref": "#/components/parameters/MaxProjects" },
          { "$ref": "#/components/parameters/RelativeTime" },
          { "$ref": "#/components/parameters/ShowSparklines" },
          { "$ref": "#/components/parameters/ShowBorder" },
          { "$ref": "#/components/parameters/Format" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/SVGCard" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/modrinth/user/{username}/downloads": {
      "get": {
        "tags": ["Modrinth"],
        "summary": "Get Modrinth user total downloads badge",
        "parameters": [
          { "$ref": "#/components/parameters/Username" },
          { "$ref": "#/components/parameters/Color" },
          { "$ref": "#/components/parameters/BackgroundColor" },
          { "$ref": "#/components/parameters/Format" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/SVGBadge" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/modrinth/user/{username}/projects": {
      "get": {
        "tags": ["Modrinth"],
        "summary": "Get Modrinth user projects count badge",
        "parameters": [
          { "$ref": "#/components/parameters/Username" },
          { "$ref": "#/components/parameters/Color" },
          { "$ref": "#/components/parameters/BackgroundColor" },
          { "$ref": "#/components/parameters/Format" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/SVGBadge" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/modrinth/user/{username}/followers": {
      "get": {
        "tags": ["Modrinth"],
        "summary": "Get Modrinth user followers badge",
        "parameters": [
          { "$ref": "#/components/parameters/Username" },
          { "$ref": "#/components/parameters/Color" },
          { "$ref": "#/components/parameters/BackgroundColor" },
          { "$ref": "#/components/parameters/Format" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/SVGBadge" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/modrinth/collection/{id}": {
      "get": {
        "tags": ["Modrinth"],
        "summary": "Get Modrinth collection card",
        "description": "Returns a rich SVG card showing collection stats and projects.",
        "parameters": [
          { "$ref": "#/components/parameters/Id" },
          { "$ref": "#/components/parameters/Color" },
          { "$ref": "#/components/parameters/BackgroundColor" },
          { "$ref": "#/components/parameters/ShowProjects" },
          { "$ref": "#/components/parameters/MaxProjects" },
          { "$ref": "#/components/parameters/RelativeTime" },
          { "$ref": "#/components/parameters/ShowSparklines" },
          { "$ref": "#/components/parameters/ShowBorder" },
          { "$ref": "#/components/parameters/Format" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/SVGCard" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/modrinth/collection/{id}/downloads": {
      "get": {
        "tags": ["Modrinth"],
        "summary": "Get Modrinth collection total downloads badge",
        "parameters": [
          { "$ref": "#/components/parameters/Id" },
          { "$ref": "#/components/parameters/Color" },
          { "$ref": "#/components/parameters/BackgroundColor" },
          { "$ref": "#/components/parameters/Format" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/SVGBadge" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/modrinth/collection/{id}/projects": {
      "get": {
        "tags": ["Modrinth"],
        "summary": "Get Modrinth collection projects count badge",
        "parameters": [
          { "$ref": "#/components/parameters/Id" },
          { "$ref": "#/components/parameters/Color" },
          { "$ref": "#/components/parameters/BackgroundColor" },
          { "$ref": "#/components/parameters/Format" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/SVGBadge" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/modrinth/collection/{id}/followers": {
      "get": {
        "tags": ["Modrinth"],
        "summary": "Get Modrinth collection followers badge",
        "parameters": [
          { "$ref": "#/components/parameters/Id" },
          { "$ref": "#/components/parameters/Color" },
          { "$ref": "#/components/parameters/BackgroundColor" },
          { "$ref": "#/components/parameters/Format" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/SVGBadge" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/curseforge/user/{id}": {
      "get": {
        "tags": ["CurseForge"],
        "summary": "Get CurseForge user card",
        "description": "Returns a rich SVG card showing user stats and top projects.",
        "parameters": [
          { "$ref": "#/components/parameters/Id" },
          { "$ref": "#/components/parameters/Color" },
          { "$ref": "#/components/parameters/BackgroundColor" },
          { "$ref": "#/components/parameters/ShowProjects" },
          { "$ref": "#/components/parameters/MaxProjects" },
          { "$ref": "#/components/parameters/RelativeTime" },
          { "$ref": "#/components/parameters/ShowSparklines" },
          { "$ref": "#/components/parameters/ShowBorder" },
          { "$ref": "#/components/parameters/Format" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/SVGCard" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/curseforge/lookup/user/{username}": {
      "get": {
        "tags": ["CurseForge"],
        "summary": "Lookup CurseForge user ID by username",
        "description": "Resolves a CurseForge username to its numeric ID by searching for their projects.",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "CurseForge username",
            "example": "gecko"
          }
        ],
        "responses": {
          "200": {
            "description": "User ID found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "description": "CurseForge user ID" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/hangar/user/{username}": {
      "get": {
        "tags": ["Hangar"],
        "summary": "Get Hangar user card",
        "description": "Returns a rich SVG card showing user stats and top projects.",
        "parameters": [
          { "$ref": "#/components/parameters/Username" },
          { "$ref": "#/components/parameters/Color" },
          { "$ref": "#/components/parameters/BackgroundColor" },
          { "$ref": "#/components/parameters/ShowProjects" },
          { "$ref": "#/components/parameters/MaxProjects" },
          { "$ref": "#/components/parameters/RelativeTime" },
          { "$ref": "#/components/parameters/ShowSparklines" },
          { "$ref": "#/components/parameters/ShowBorder" },
          { "$ref": "#/components/parameters/Format" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/SVGCard" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "Username": {
        "name": "username",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Username",
        "example": "prospector"
      },
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Entity ID",
        "example": "VEgGDFFE"
      },
...
      "Color": {
        "name": "color",
        "in": "query",
        "schema": { "type": "string", "pattern": "^#[0-9A-Fa-f]{6}$" },
        "description": "Accent color in hex format. Defaults to platform brand color."
      },
      "BackgroundColor": {
        "name": "backgroundColor",
        "in": "query",
        "schema": { "type": "string", "default": "transparent" },
        "description": "Background color in hex format or 'transparent'"
      },
      "ShowProjects": {
        "name": "showProjects",
        "in": "query",
        "schema": { "type": "boolean", "default": true },
        "description": "Show top projects section"
      },
...
      "MaxProjects": {
        "name": "maxProjects",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "maximum": 10, "default": 5 },
        "description": "Maximum projects to display (1-10)"
      },
...
      "RelativeTime": {
        "name": "relativeTime",
        "in": "query",
        "schema": { "type": "boolean", "default": true },
        "description": "Show relative time for dates (e.g., '2 days ago')"
      },
      "ShowSparklines": {
        "name": "showSparklines",
        "in": "query",
        "schema": { "type": "boolean", "default": true },
        "description": "Display activity sparklines"
      },
      "ShowBorder": {
        "name": "showBorder",
        "in": "query",
        "schema": { "type": "boolean", "default": true },
        "description": "Show outer border on the card"
      },
      "Format": {
        "name": "format",
        "in": "query",
        "schema": { "type": "string", "enum": ["svg", "png"], "default": "svg" },
        "description": "Output format"
      }
    },
    "responses": {
      "SVGCard": {
        "description": "SVG card image",
        "content": {
          "image/svg+xml": {
            "schema": { "type": "string", "format": "binary" }
          }
        }
      },
      "SVGBadge": {
        "description": "SVG badge image",
        "content": {
          "image/svg+xml": {
            "schema": { "type": "string", "format": "binary" }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" }
              }
            }
          }
        }
      }
    }
  }
}